:recycle: refactor service code with lombok
