+++ src/main/java/com/saltynote/service/service/NoteService.java
@@ -1,15 +1,13 @@
 package com.saltynote.service.service;
 
 import com.saltynote.service.repository.NoteRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class NoteService implements RepositoryService<NoteRepository> {
     private final NoteRepository repository;
-
-    public NoteService(NoteRepository repository) {
-        this.repository = repository;
-    }
 
     @Override
     public NoteRepository getRepository() {
