+++ src/main/java/com/saltynote/service/service/UserDetailsServiceImpl.java
@@ -2,16 +2,16 @@ package com.saltynote.service.service;
 
 import com.saltynote.service.domain.LoginUser;
 import com.saltynote.service.entity.SiteUser;
-import jakarta.annotation.Resource;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class UserDetailsServiceImpl implements UserDetailsService {
 
-    @Resource
-    private UserService userService;
+    private final UserService userService;
 
     @Override
     public LoginUser loadUserByUsername(String username) throws UsernameNotFoundException {
